+++ packages/evmos-wallet/src/internal/wallet/functionality/errors.ts
@@ -44,6 +44,8 @@ export const METAMASK_NOTIFICATIONS = {
   SuccessTitle: "Connected with Metamask",
   PubkeySubtext:
     "Could not get the pubkey, please sign the generate_pubkey message!",
+  ExtensionNotFoundSubtext:
+    "The MetaMask extension could not be found. Try using Copilot to get started!",
   ChangeNetworkSubtext: "Could not change the network to EVMOS",
   AddressSubtext: "Could not get the user address from Metamask",
   DeniedSignatureSubtext: "Transaction was not signed",
@@ -57,7 +59,8 @@ export const METAMASK_NOTIFICATIONS = {
 export const KEPLR_NOTIFICATIONS = {
   ErrorTitle: "Error connecting with Keplr",
   SuccessTitle: "Connected with Keplr",
-  ExtensionNotFoundSubtext: "The Keplr extension could not be found",
+  ExtensionNotFoundSubtext:
+    "The Keplr extension could not be found. Try using Copilot to get started!",
   LedgerNotInitSubtext: "Please initialize ethereum app on ledger first",
   RequestRejectedSubtext:
     "Please unlock the extension and allow the app to access your wallet address",