+++ packages/evmos-wallet/src/wallet/ButtonWalletConnection.tsx
@@ -43,9 +43,9 @@ import {
 } from "../internal/wallet/functionality/walletconnect/walletconnect";
 import { Tooltip } from "ui-helpers";
 import {
-  CLICK_WC_CONNECT_WALLET_BUTTON,
-  CLICK_WC_DISCONNECT_WALLET_BUTTON,
-  CLICK_WC_CONNECTED_WITH,
+  CLICK_CONNECT_WALLET_BUTTON,
+  CLICK_DISCONNECT_WALLET_BUTTON,
+  CLICK_CONNECTED_WITH,
   SWITCH_BETWEEN_WALLETS,
   SUCCESSFUL_WALLET_CONNECTION,
   UNSUCCESSFUL_WALLET_CONNECTION,
@@ -53,9 +53,11 @@ import {
 } from "tracker";
 // Components
 import { Button } from "ui-helpers";
+import { ProvidersIcons } from "../copilot/utils";
 import { TranslationContextProvider } from "schummar-translate/react";
 import { t } from "../locales/translate";
 
+// TODO: remove it. Deprecated. Now we are using copilot -> WalletConnection
 export const ButtonWalletConnection = ({
   walletExtension,
   dispatch,
@@ -65,7 +67,7 @@ export const ButtonWalletConnection = ({
 }) => {
   const [show, setShow] = useState(false);
 
-  const useWC = useWalletConnect(store);
+  const useWalletConnectHook = useWalletConnect(store);
 
   const close = useCallback(() => setShow(false), []);
   const open = useCallback(() => setShow(true), []);
@@ -108,15 +110,14 @@ export const ButtonWalletConnection = ({
   const [isCopied, setIsCopied] = useState(false);
 
   const { handlePreClickAction: trackClickConnectWallet } = useTracker(
-    CLICK_WC_CONNECT_WALLET_BUTTON
+    CLICK_CONNECT_WALLET_BUTTON
   );
   const { handlePreClickAction: trackClickDisconnectWallet } = useTracker(
-    CLICK_WC_DISCONNECT_WALLET_BUTTON
+    CLICK_DISCONNECT_WALLET_BUTTON
   );
 
-  const { handlePreClickAction: trackConnectedWithWallet } = useTracker(
-    CLICK_WC_CONNECTED_WITH
-  );
+  const { handlePreClickAction: trackConnectedWithWallet } =
+    useTracker(CLICK_CONNECTED_WITH);
 
   const { handlePreClickAction: trackChangeWallet } = useTracker(
     SWITCH_BETWEEN_WALLETS
@@ -178,12 +179,7 @@ export const ButtonWalletConnection = ({
             className="flex items-center space-x-3 justify-center"
             onClick={open}
           >
-            {walletExtension.extensionName === METAMASK_KEY && <MetamaskIcon />}
-            {walletExtension.extensionName === KEPLR_KEY && <KeplrIcon />}
-            {walletExtension.extensionName === WALLECT_CONNECT_KEY && (
-              <WalletConnectIcon />
-            )}
-
+            {ProvidersIcons[walletExtension.extensionName]}
             <span className="text-lg font-bold">
               {formatProviderAddress(walletExtension, true)}
             </span>
@@ -342,7 +338,7 @@ export const ButtonWalletConnection = ({
                 <ButtonWallet
                   onClick={async () => {
                     setShow(false);
-                    await useWC.connect();
+                    await useWalletConnectHook.connect();
                     trackConnectedWithWallet({
                       wallet: GetWalletFromLocalStorage(),
                       provider: WALLECT_CONNECT_KEY,