+++ packages/evmos-wallet/src/wallet/ContentModalConnect.tsx
@@ -1,9 +1,9 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
 
-const ContentModalConnect = ({ children }: { children: JSX.Element }) => {
+const ContentModalConnect = ({ children }: { children: React.ReactNode }) => {
   return (
-    <div className="flex items-center space-x-5 text-lg font-bold">
+    <div className="flex items-center space-x-3 text-sm font-bold">
       {children}
     </div>
   );